refactor (ontology cache): fix circular deps problem
