+++ src/index.ts
@@ -14,7 +14,6 @@ export { LoginResponse } from "./models/v2/authentication/login-response";
 export { LogoutResponse } from "./models/v2/authentication/logout-response";
 
 export { UserCache } from "./cache/UserCache";
-export { OntologyCache } from "./cache/OntologyCache";
 export { ListNodeCache } from "./cache/ListNodeCache";
 
 export { ApiResponse } from "./models/api-response";
