+++ src/models/v2/resources/ResourcesConversionUtil.ts
@@ -1,8 +1,8 @@
 import { JsonConvert } from "json2typescript";
 import { forkJoin, Observable, of } from "rxjs";
 import { map, mergeMap } from "rxjs/operators";
-import { ListNodeCache, OntologyCache } from "../../../";
-import { IResourceClassAndPropertyDefinitions } from "../../../cache/OntologyCache";
+import { ListNodeCache } from "../../../";
+import { IResourceClassAndPropertyDefinitions, OntologyCache } from "../../../cache/OntologyCache";
 import { Constants } from "../Constants";
 import { ResourcePropertyDefinition } from "../ontologies/resource-property-definition";
 import { CountQueryResponse } from "../search/count-query-response";
